IOT-2204:  Fix IPCA compiler warnings on Linux
https://jira.iotivity.org/browse/IOT-2204

Reviewed-on: https://gerrit.iotivity.org/gerrit/19753
Tested-by: jenkins-iotivity <jenkins@iotivity.org>

+++ resource/IPCA/src/inc/ocfframework.h
@@ -39,7 +39,7 @@ typedef struct DeviceDetails
     bool deviceNotRespondingIndicated;
 
     // How many IPCAOpenDevice() on this device.
-    size_t deviceOpenCount;
+    int deviceOpenCount;
 
     // Timestamp of final close.
     uint64_t lastCloseDeviceTime;
